% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exif.R
\docType{package}
\name{exif}
\alias{exif}
\alias{exif-package}
\title{Read EXIF data into R}
\description{
exif is a package for reading EXIF media metadata into R,
returning it as a list in a similar fashion to jsonlite. It depends on
the libexif C library, which must be installed for the package to work.
}
\seealso{
\code{\link{read_exif}}
}
\author{
\strong{Maintainer}: Os Keyes \email{ironholds@gmail.com}

Other contributors:
\itemize{
  \item Penelope Hopkins [contributor]
  \item Robyn Temple-Wood [contributor]
  \item Mayank Lahiri [copyright holder]
}

}
