% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_exif}
\alias{read_exif}
\title{read EXIF data into R}
\usage{
read_exif(filenames)
}
\arguments{
\item{filenames}{a vector of full filenames.}
}
\value{
A nested set of lists, containing
the metadata for each file. If only one file
is requested, the metadata will be at the "top level"
of the list, rather than unnecessarily nested; if
a file cannot be found or read correctly, the metadata
for that file will consist of the line "Could not read file".
}
\description{
\code{read_exif}
}

